+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-common/src/main/java/org/apache/hadoop/yarn/api/records/impl/pb/ContainerPBImpl.java
@@ -206,7 +206,7 @@ public class ContainerPBImpl extends Container {
   }
 
   @Override
-  public void setGPULocation(int GPULocation) {
+  public void setGPULocation(long GPULocation) {
     maybeInitBuilder();
     if (resource == null)
       builder.clearResource();
